feat: implement style extraction utility and enhance table processing with improved cell styling
